requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
numpy>=1.21.0

# Image processing
opencv-python>=4.8.0
Pillow>=9.5.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web frameworks
fastapi>=0.100.0
uvicorn>=0.22.0
streamlit>=1.25.0

# Data handling
pandas>=2.0.0
pyyaml>=6.0

# Plotting and visualization
plotly>=5.15.0

# Utilities
tqdm>=4.65.0
pathlib2>=2.3.7

albumentations
scikit-image
